// src/app/enums/game-codes.ts
export enum GameStatus {
    NOT_STARTED = "Not Started",
    STARTED = "Started",
    PAUSED = "Paused",
    GAME_OVER = "Game Over",
    WON = "Won"
}

export enum MoveDirection {
    LEFT_TO_RIGHT = "LR",
    RIGHT_TO_LEFT = "RL",
    TOP_TO_BOTTOM = "TB",
    BOTTOM_TO_TOP = "BT"
}

export enum MovableItems {
    BIRD = "Bird",
    EAGLE = "Eagle",
    SCORPION = "Scorpion",
    SNIPER = "Sniper"
}